Work on getting Tailwind CSS to Work
Currently Tailwind CSS does not work.
