+++ src/app/components/Common/Container/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Container';
