--- src/app/components/CourseIdentifierAutocomplete/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./CourseIdentifierAutocomplete";
