+++ src/app/components/CourseSelector/CourseSelectorAutocomplete/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CourseSelectorAutocomplete";
