+++ tailwind.config.ts
@@ -1,6 +1,8 @@
 import type { Config } from "tailwindcss";
+import themeConfig from "./src/assets/config/theme.json";
 
 const config: Config = {
+  darkMode: ["class", '[data-mantine-color-scheme="dark"]'],
   content: [
     "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
@@ -8,9 +10,20 @@ const config: Config = {
   ],
   theme: {
     extend: {
+      screens: {
+        xsm: "400px",
+      },
       colors: {
-        background: "var(--background)",
-        foreground: "var(--foreground)",
+        "primary-color": themeConfig.color.primary.base,
+        "primary-color-light": themeConfig.color.primary.light,
+        "secondary-color-light": themeConfig.color.secondary.light,
+        "secondary-color": themeConfig.color.secondary.base,
+        "secondary-color-dark": themeConfig.color.secondary.dark,
+        "tertiary-color": themeConfig.color.tertiary.base,
+        "tertiary-color-dark": themeConfig.color.tertiary.dark,
+      },
+      height: {
+        dscreen: "100dvh", // Use 100dvh insteda of 100vh, so that it works on mobile.
       },
     },
   },
